% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gexp}
\alias{gexp}
\title{Single Gene Expression Data}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
gexp
}
\description{
A simple gene expression data in two groups. (Based on the \code{sleep}
data).
}
\examples{
data(gexp)
gexp
}
\keyword{datasets}
