% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmdata}
\alias{lmdata}
\alias{lmannot}
\title{Gene expression data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 12 columns.
}
\usage{
lmdata
}
\description{
Factorial design (2 factors with 2 levels each) with gene expression for 5
genes and 12 samples.
}
\examples{
data(lmdata, verbose = TRUE)
lmdata
lmannot
}
\keyword{datasets}
