% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{prepare_shell}
\alias{prepare_shell}
\title{Prepare Shell data}
\usage{
prepare_shell(shell_dir = "wsbim2122_data", rm_dir = FALSE)
}
\arguments{
\item{shell_dir}{\code{character(1)} with the name of the shell data directory. Default
is \code{"wsbim2122_data"}.}

\item{rm_dir}{\code{logical(1)} that defines whether \code{shell_dir} is going to be
recursively removed. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
A function used for its side effect. It creates the directory and files for
the shell lesson.
}
\author{
Laurent Gatto
}
