% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWSBIM2122.R
\name{rWSBIM2122version}
\alias{rWSBIM2122version}
\title{Package version}
\usage{
rWSBIM2122version()
}
\description{
Package version
}
\examples{
## check the package version that is currently installed
rWSBIM2122version()
}
